requirements.txt
annotated-types==0.6.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.2
Flask-Cors==4.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
setuptools==68.2.2
typing_extensions==4.10.0
Werkzeug==3.0.1
wheel==0.41.2
